Removed MSVC 2010 warnings

+++ CORE/edge.cpp
@@ -42,7 +42,7 @@ void Edge::computeMargin()
 {
 	if (_fontSize) {
 		Coordinates tm(
-		  _size + ceil(AVG_CHAR_WIDTH * _fontSize * _text.length()),
+		  _size + (int)ceil(AVG_CHAR_WIDTH * _fontSize * _text.length()),
 		  _size / 2 + _fontSize
 		);
 		_margin.setRb(tm);
